+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 
 - Added task content interface to differentiate from behavior
   [sgeulette]
+- Added task_description field
+  [sgeulette]
 - Added assigned_group and due_date indexes. Added indexer methods avoiding acquisition for children. Do not store None in catalog
   [sgeulette]
 - Changed workflow to use 'Request review' guard permission to differentiate Reviewer and editor transitions (like in plone workflows)